add balance

+++ internal/broker/model/order.go
@@ -8,5 +8,5 @@ type Order struct {
 type OrderAccrual struct {
 	Order   string  `json:"order"`
 	Status  string  `json:"status"`
-	Accrual float64 `json:"accrual,omitempty"`
+	Accrual float32 `json:"accrual,omitempty"`
 }
